Change from using boost::ptr_map to storing ANC,DESC in vertices
MAJOR UPDATE. This had to be done because boost::ptr_map was
all weird and couldn't figure out why. Sometime master branch will
have to be updated to and there will be two versions, this one
with the built-in Graph type and the other that should be generic.
IN THIS COMMIT ALSO: use namespaces boost and std, change variable
names to shorten code a bit, and fix data str init with correct
data types.